adding checks for no files in the bucket and for empty json responses

+++ upload_customfields.py
@@ -9,6 +9,10 @@ def get_file_paths(s3_client, bucket_name, prefix):
     print('\t getting all file_paths')
     bucket_contents = s3_client.list_objects(Bucket=bucket_name, Prefix=prefix)
 
+    if not 'Contents' in bucket_contents.keys(): # no files in the bucket
+        print('\t\t no files found in bucket path')
+        return False
+
     file_paths = [x['Key'] for x in bucket_contents['Contents']]
 
     return file_paths
@@ -58,30 +62,33 @@ def main():
 
     file_paths = get_file_paths(s3_client, bucket_name, prefix='data/customfields/')
 
-    for file_path in file_paths:
-        json_file = get_json_file(s3_client, bucket_name, file_path)
-        request_date = file_path[18:28]
-        customfields = create_dataframe(json_file)
+    if file_paths: # files found, no problem
+        for file_path in file_paths:
+            json_file = get_json_file(s3_client, bucket_name, file_path)
+            request_date = file_path[18:28]
 
-        template = ', '.join(['%s'] * len(customfields.columns))
-        query = '''INSERT INTO customfields
-                      (active, applies_to, created, last_modified, name,
-                      regex_filter, required, required_customfields, short_code,
-                      type, ui_preference, customfield_id, last_updated)
-                   VALUES ({template}, '{last_updated}')
-                   ON CONFLICT(customfield_id) DO
-                   UPDATE SET
-                      active = excluded.active, applies_to = excluded.applies_to,
-                      created = excluded.created, last_modified = excluded.last_modified,
-                      name = excluded.name, regex_filter = excluded.regex_filter,
-                      required = excluded.required, required_customfields = excluded.required_customfields,
-                      short_code = excluded.short_code, type = excluded.type,
-                      ui_preference = excluded.ui_preference, customfield_id = excluded.customfield_id,
-                      last_updated = excluded.last_updated'''.format(template=template, last_updated=str(request_date))
+            if json_file['results']['customfields'] != []:
+                customfields = create_dataframe(json_file)
 
-        upload_to_db(conn, customfields, query)
+                template = ', '.join(['%s'] * len(customfields.columns))
+                query = '''INSERT INTO customfields
+                              (active, applies_to, created, last_modified, name,
+                              regex_filter, required, required_customfields, short_code,
+                              type, ui_preference, customfield_id, last_updated)
+                           VALUES ({template}, '{last_updated}')
+                           ON CONFLICT(customfield_id) DO
+                           UPDATE SET
+                              active = excluded.active, applies_to = excluded.applies_to,
+                              created = excluded.created, last_modified = excluded.last_modified,
+                              name = excluded.name, regex_filter = excluded.regex_filter,
+                              required = excluded.required, required_customfields = excluded.required_customfields,
+                              short_code = excluded.short_code, type = excluded.type,
+                              ui_preference = excluded.ui_preference, customfield_id = excluded.customfield_id,
+                              last_updated = excluded.last_updated'''.format(template=template, last_updated=str(request_date))
 
-        s3_client.delete_object(Bucket=bucket_name, Key=file_path)
+                upload_to_db(conn, customfields, query)
+
+            s3_client.delete_object(Bucket=bucket_name, Key=file_path)
 
     conn.close()
