+++ upload_employees.py
@@ -9,6 +9,9 @@ def get_file_paths(s3_client, bucket_name, prefix):
     print('\t getting all file_paths')
     bucket_contents = s3_client.list_objects(Bucket=bucket_name, Prefix=prefix)
 
+    if not 'Contents' in bucket_contents.keys(): # no files in the bucket
+        print('\t\t no files found in bucket path')
+        return False
     file_paths = [x['Key'] for x in bucket_contents['Contents']]
 
     return file_paths
@@ -66,43 +69,46 @@ def main():
 
     file_paths = get_file_paths(s3_client, bucket_name, prefix='data/employees/')
 
-    for file_path in file_paths:
-        json_file = get_json_file(s3_client, bucket_name, file_path)
-        request_date = file_path[15:25]
-        employees = create_dataframe(json_file)
+    if file_paths: # file paths found, continue
+        for file_path in file_paths:
+            json_file = get_json_file(s3_client, bucket_name, file_path)
+            request_date = file_path[15:25]
 
-        template = ', '.join(['%s'] * len(employees.columns))
-        query = '''INSERT INTO employees
-                      (active, approved_to, client_url, company_name, created,
-                      customfields, email, email_verified, employee_number, exempt,
-                      first_name, group_id, hire_date, last_active, last_modified,
-                      last_name, manager_of_group_ids, mobile_number, pay_interval,
-                      pay_rate, payroll_id, permissions, profile_image_url,
-                      pto_balances, require_password_change, salaried, submitted_to,
-                      term_date, username, employee_id, last_updated)
-                   VALUES ({template}, '{last_updated}')
-                   ON CONFLICT(employee_id) DO
-                   UPDATE SET
-                       active = excluded.active, approved_to = excluded.approved_to,
-                       client_url = excluded.client_url, company_name = excluded.company_name,
-                       created = excluded.created, customfields = excluded.customfields,
-                       email = excluded.email, email_verified = excluded.email_verified,
-                       employee_number = excluded.employee_number, exempt = excluded.exempt,
-                       first_name = excluded.first_name, group_id = excluded.group_id,
-                       hire_date = excluded.hire_date, last_active = excluded.last_active,
-                       last_modified = excluded.last_modified, last_name = excluded.last_name,
-                       manager_of_group_ids = excluded.manager_of_group_ids,
-                       mobile_number = excluded.mobile_number, pay_interval = excluded.pay_interval,
-                       pay_rate = excluded.pay_rate, payroll_id = excluded.payroll_id,
-                       permissions = excluded.permissions, profile_image_url = excluded.profile_image_url,
-                       pto_balances = excluded.pto_balances, require_password_change = excluded.require_password_change,
-                       salaried = excluded.salaried, submitted_to = excluded.submitted_to,
-                       term_date = excluded.term_date, username = excluded.username,
-                       employee_id = excluded.employee_id, last_updated = excluded.last_updated'''.format(template=template, last_updated=str(request_date))
+            if json_file['results']['users'] != []: # check for empty response
+                employees = create_dataframe(json_file)
 
-        upload_to_db(conn, employees, query)
+                template = ', '.join(['%s'] * len(employees.columns))
+                query = '''INSERT INTO employees
+                              (active, approved_to, client_url, company_name, created,
+                              customfields, email, email_verified, employee_number, exempt,
+                              first_name, group_id, hire_date, last_active, last_modified,
+                              last_name, manager_of_group_ids, mobile_number, pay_interval,
+                              pay_rate, payroll_id, permissions, profile_image_url,
+                              pto_balances, require_password_change, salaried, submitted_to,
+                              term_date, username, employee_id, last_updated)
+                           VALUES ({template}, '{last_updated}')
+                           ON CONFLICT(employee_id) DO
+                           UPDATE SET
+                               active = excluded.active, approved_to = excluded.approved_to,
+                               client_url = excluded.client_url, company_name = excluded.company_name,
+                               created = excluded.created, customfields = excluded.customfields,
+                               email = excluded.email, email_verified = excluded.email_verified,
+                               employee_number = excluded.employee_number, exempt = excluded.exempt,
+                               first_name = excluded.first_name, group_id = excluded.group_id,
+                               hire_date = excluded.hire_date, last_active = excluded.last_active,
+                               last_modified = excluded.last_modified, last_name = excluded.last_name,
+                               manager_of_group_ids = excluded.manager_of_group_ids,
+                               mobile_number = excluded.mobile_number, pay_interval = excluded.pay_interval,
+                               pay_rate = excluded.pay_rate, payroll_id = excluded.payroll_id,
+                               permissions = excluded.permissions, profile_image_url = excluded.profile_image_url,
+                               pto_balances = excluded.pto_balances, require_password_change = excluded.require_password_change,
+                               salaried = excluded.salaried, submitted_to = excluded.submitted_to,
+                               term_date = excluded.term_date, username = excluded.username,
+                               employee_id = excluded.employee_id, last_updated = excluded.last_updated'''.format(template=template, last_updated=str(request_date))
 
-        s3_client.delete_object(Bucket=bucket_name, Key=file_path)
+                upload_to_db(conn, employees, query)
+
+            s3_client.delete_object(Bucket=bucket_name, Key=file_path)
 
     conn.close()
